+++ isitfit/__init__.py
@@ -1,1 +1,1 @@
-isitfit_version='0.5.2'
+isitfit_version='0.5.3'