+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version latest (0.5.0?, 2019-09-24?)
 
 - ...
+- enh: separate consolidation of all results from displaying the consolidation
 
 
 Version 0.5.2 (2019-09-25)