+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version latest (0.5.0?, 2019-09-24?)
 - ...
 - enh: separate consolidation of all results from displaying the consolidation
 - feat: stream intermediate results to csv during the optimization
+- feat: save optimization final output to csv
 
 
 Version 0.5.2 (2019-09-25)